(#168) Refactoring
